Created library for easy Azure function creation
cosmotech_scenario_download.generate_main.generate_main is a function allowing fast creation of Azure function.
It takes a single function as an argument and will apply it to the results of the scenario_downloader.

Base ScenarioDownload azure function is left as a simple example of usage of the new library